% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{check_demog_file}
\alias{check_demog_file}
\title{check_demog_file}
\usage{
check_demog_file(demog_list, pars_list, check = FALSE)
}
\arguments{
\item{demog_list}{Output of \link{samrat_read_demography}}
}
\description{
\code{check_demog_file} Checks demog file is correctly structured
}
\keyword{internal}
