% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{check_param_file}
\alias{check_param_file}
\title{check_param_file}
\usage{
check_param_file(pars_list)
}
\arguments{
\item{pars_list}{Output of \link{samrat_read_params}}
}
\description{
\code{check_param_file} Checks param file is correctly structures
}
\keyword{internal}
