% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{check_predictors_file}
\alias{check_predictors_file}
\title{check_predictors_file}
\usage{
check_predictors_file(predictors_list, pars_list, check = FALSE)
}
\arguments{
\item{predictors_list}{Output of \link{samrat_read_predictors}}
}
\description{
\code{check_predictors_file} Checks demog file is correctly structured
}
\keyword{internal}
