% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{f_calculate_folds}
\alias{f_calculate_folds}
\title{Calculate folds}
\usage{
f_calculate_folds(training_data, admin2_col, k_folds = 10)
}
\arguments{
\item{training_data}{Training Dataset}

\item{admin2_col}{Colnames of admin2}
}
\description{
Calculate folds
}
\details{
Calculate folds
}
