% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_retrospective.R
\name{f_cdr_map}
\alias{f_cdr_map}
\title{Function to produce plot for CDR at district level}
\usage{
f_cdr_map(
  data,
  resp_var,
  dir_output = "08_define_final_model/visualisation/output"
)
}
\arguments{
\item{data}{Responsible results}

\item{resp_var}{responsive variable}

\item{dir_output}{folder to save the output}
}
\description{
Function to produce plot for CDR at district level
}
\details{
Function to produce plot for CDR at district level
}
