% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_counterfactuals.R
\name{f_clean_actual_table}
\alias{f_clean_actual_table}
\title{Function to return clean actual counterfactuals table}
\usage{
f_clean_actual_table(actual_res)
}
\arguments{
\item{actual_res}{Actual Results}
}
\description{
Function to return clean actual counterfactuals table
}
\details{
Function to return clean actual counterfactuals table
}
