% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_counterfactuals.R
\name{f_clean_counterfactuals_table}
\alias{f_clean_counterfactuals_table}
\title{Function to return clean excess counterfactuals table}
\usage{
f_clean_counterfactuals_table(excess_res)
}
\arguments{
\item{excess_res}{Excess Results}
}
\description{
Function to return clean excess counterfactuals table
}
\details{
Function to return clean counterfactuals table
}
