% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual.R
\name{f_counterfactuals}
\alias{f_counterfactuals}
\title{Function to generate counterfactuals}
\usage{
f_counterfactuals(
  boostrapping_results,
  resp_var,
  agg_level,
  admin2_col,
  start_date_per_1 = "2015-01-01",
  end_date_per_1 = "2016-06-01",
  start_date_per_2 = "2022-01-01",
  end_date_per_2 = "2022-12-01",
  nb_boostrap = 1000,
  type_model = "glm"
)
}
\arguments{
\item{boostrapping_results}{Datasets containing boostraping}

\item{resp_var}{response variable}

\item{agg_level}{List of levels to aggregate}

\item{admin2_col}{colnames of admin2}

\item{start_date_per_1}{Start date period 1}

\item{end_date_per_1}{End date period 1}

\item{start_date_per_2}{Start date period 2}

\item{end_date_per_2}{End date period 2}

\item{nb_boostrap}{number of boostrapping}

\item{type_model}{type of model}
}
\description{
Function to generate counterfactuals
}
\details{
Function to generate counterfactuals
}
