% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_retrospective.R
\name{f_country_dr_plot}
\alias{f_country_dr_plot}
\title{Function to plot countrywide death rate evolution}
\usage{
f_country_dr_plot(
  res_data,
  resp_var = "cdr",
  dir_output = "08_define_final_model/visualisation/output"
)
}
\arguments{
\item{res_data}{Responsible results}

\item{resp_var}{responsive variable}

\item{dir_output}{folder to save the output}
}
\description{
Function to plot countrywide death rate evolution
}
\details{
Function to plot countrywide death rate evolution
}
