% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{f_cv}
\alias{f_cv}
\title{Function to generate cross validation}
\usage{
f_cv(
  data_f,
  which_fold_f,
  fit_f,
  f_predict_f,
  f_family_f,
  k_folds_f = 10,
  verbose = TRUE
)
}
\arguments{
\item{data_f}{Datasets containing the different folds}

\item{which_fold_f}{Which folds are we using}

\item{fit_f}{model fitting function}

\item{f_family_f}{family function of the fit (e.g poisson)}

\item{k_folds_f}{Number of folds}
}
\description{
Function to generate cross validation
}
\details{
Function to perform cross validation
}
