% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_visualisation.R
\name{f_cv_plot}
\alias{f_cv_plot}
\title{Function to generate plot of cross-validation performance by region-year}
\usage{
f_cv_plot(
  cv_data,
  admin1_col,
  plot_type = "scatter",
  dir_output_f = dir_output,
  save_plot = TRUE
)
}
\arguments{
\item{cv_data}{cross validation data}

\item{admin1_col}{Colnames of admin1}

\item{plot_type}{scatter or dumbbell}

\item{dir_output_f}{direction output to save the dataset}

\item{save_plot}{save plot True or False}

\item{which_fold_f}{Name of the fold used}
}
\description{
Function to generate plot of cross-validation performance by region-year
}
\details{
Function to generate plot of cross-validation performance by region-year
}
