% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{f_define_annot_value}
\alias{f_define_annot_value}
\title{Define annot variable}
\usage{
f_define_annot_value(resp_var)
}
\arguments{
\item{resp_var}{responsive variable}
}
\description{
Define annot variable
}
\details{
Define annot variable
}
