% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{f_define_p_time}
\alias{f_define_p_time}
\title{Define p_time variable}
\usage{
f_define_p_time(resp_var)
}
\arguments{
\item{resp_var}{responsive variable}
}
\description{
Define p_time variable
}
\details{
Define p_time variable
}
