% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{f_define_pop_var}
\alias{f_define_pop_var}
\title{Define pop_var variable}
\usage{
f_define_pop_var(resp_var)
}
\arguments{
\item{resp_var}{responsive variable}
}
\description{
Define pop_var variable
}
\details{
Define pop variable
}
