% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{f_define_pred_value}
\alias{f_define_pred_value}
\title{Define pred values variable}
\usage{
f_define_pred_value(resp_var)
}
\arguments{
\item{resp_var}{responsive variable}
}
\description{
Define pred values variable
}
\details{
Define pred_value variable
}
