% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{f_family}
\alias{f_family}
\title{Function to extract the distribution ("family") used to fit the model}
\usage{
f_family(fit_f = fit)
}
\arguments{
\item{fit_f}{model fitting function}
}
\description{
Function to extract the distribution ("family") used to fit the model
}
\details{
Function to extract the distribution ("family") used to fit the model
}
