% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{f_feature_selection}
\alias{f_feature_selection}
\title{Function to generate feature selection}
\usage{
f_feature_selection(
  list_pred_to_try,
  pred_value,
  path,
  training_data,
  which_fold_f,
  admin2_col = NULL,
  p_time_var = NULL,
  model = "glmmTMB",
  admin1_col = "region"
)
}
\arguments{
\item{list_pred_to_try}{list of predictor variables to try}

\item{pred_value}{value we want to predict}

\item{path}{path to save the dataset}

\item{training_data}{training dataset}

\item{which_fold_f}{Name of the fold used}

\item{admin2_col}{colnames of the admin2, if not used NULL}

\item{p_time_var}{p_time_var used or not (e.g NULL or p_time or p_time_u5)}

\item{model}{statistic model (e.g. glmmTB)}
}
\description{
Function to generate feature selection
}
\details{
Function to perform cross validation
}
