% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_and_boostraps.R
\name{f_generate_final_results}
\alias{f_generate_final_results}
\title{Generate the actual mean, low and upp dr and toll using a agg level}
\usage{
f_generate_final_results(
  boostrapping_results,
  agg_level,
  nb_boostrap,
  resp_var
)
}
\arguments{
\item{boostrapping_results}{Boostraping results}

\item{agg_level}{Aggregate levels}

\item{nb_boostrap}{number of boostrapings}

\item{resp_var}{response variable}
}
\description{
Generate the actual mean, low and upp dr and toll using a agg level
}
\details{
Generate the actual mean, low and upp dr and toll using a agg level
}
