% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smals.R
\name{f_liv}
\alias{f_liv}
\title{\code{f_liv}}
\usage{
f_liv(liv_term, liv_ss)
}
\arguments{
\item{liv_term}{Livelihood liv_term being matched with}

\item{liv_ss}{Livelihood substrings}
}
\description{
Function to standardise livelihood type nomenclature
}
\keyword{internal}
