% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_counterfactuals.R
\name{f_plot_country_plot_with_boostrap}
\alias{f_plot_country_plot_with_boostrap}
\title{Function to to plot country level curves with boostrapping}
\usage{
f_plot_country_plot_with_boostrap(
  boostrapping_results,
  name_output,
  path_output = "08_define_final_model/visualisation/output/",
  resp_var = "n_died",
  start_date_cf_a = "2015-01-01",
  end_date_cf_a = "2016-06-01",
  start_date_cf_b = "2020-01-01",
  end_date_cf_b = "2021-06-01",
  cf_b_label = "2015-16 counterfactual",
  cf_a_label = "2020 counterfactual"
)
}
\arguments{
\item{boostrapping_results}{Boostrapping results}

\item{name_output}{name of the output to save the results}

\item{path_output}{path of the folder to save the results}

\item{resp_var}{response variable}

\item{start_date_cf_a}{start date for counterfactual period 1}

\item{end_date_cf_a}{end date for counterfactual period 1}

\item{start_date_cf_b}{start date for counterfactual period 2}

\item{end_date_cf_b}{end date for counterfactual period 2}

\item{cf_b_label}{name of counterfactual period 2}

\item{cf_a_label}{name of counterfactual period 1}
}
\description{
Function to to plot country level curves with boostrapping
}
\details{
Function to to plot country level curves with boostrapping
}
