% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_and_boostraps.R
\name{f_pred_glm}
\alias{f_pred_glm}
\title{Train, predict glmmTMB and return the result with boostraps}
\usage{
f_pred_glm(
  training_data,
  predicting_data,
  pred_var,
  resp_var,
  formula = fit_formula,
  model = glm_fit,
  admin1_col = "region",
  admin2_col = "district",
  name_output = "som_glm_model"
)
}
\arguments{
\item{training_data}{Training data}

\item{predicting_data}{predicting data}

\item{resp_var}{reponse variable}

\item{admin1_col}{admin1 colnames}

\item{admin2_col}{admin2 colnames}

\item{name_output}{name to save the output}
}
\description{
Train, predict glmmTMB and return the result with boostraps
}
\details{
Train, predict glm and return the result with boostraps
}
