% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{f_predict}
\alias{f_predict}
\title{Function to predict the dataset}
\usage{
f_predict(
  fit_f,
  newdata_f,
  type_f = "response",
  resp_var = NULL,
  bootstrap_f = FALSE,
  n_bootstraps = NULL
)
}
\arguments{
\item{fit_f}{model fitting function}

\item{type_f}{type of prediction (e.g. response)}

\item{resp_var}{response variable}

\item{n_bootstraps}{number of boostrapping}

\item{new_data_f}{validation data to predict}

\item{boostrap_f}{boostrapping function}
}
\description{
Function to predict the dataset
}
\details{
Function to predict the dataset
}
\examples{
# example code

}
