% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{par_list_from_df}
\alias{par_list_from_df}
\title{par_list_from_df}
\usage{
par_list_from_df(pars, value = "value", parameter = "parameter")
}
\arguments{
\item{pars}{\code{data.frame} of parameters and values}

\item{value}{String for value column. Default = \code{"value"}}

\item{parameter}{String for value column. Default = \code{"parameter"}}
}
\description{
Convert data frame with columns value and parameter into
named list with numeric values formatted
}
\details{
\code{par_list_from_df} convert data.frame into formatted list
}
\keyword{internal}
