% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samrat.R
\docType{package}
\name{samrat}
\alias{samrat-package}
\alias{samrat}
\title{\pkg{samrat} Small-area mortality analytics}
\description{
Package for conducting small-area estimation of
mortality and short term predictions.
}
\author{
\strong{Maintainer}: OJ Watson \email{oj.watson@hotmail.co.uk} (\href{https://orcid.org/0000-0003-2374-0741}{ORCID})

}
\keyword{internal}
