% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{samrat_read_params}
\alias{samrat_read_params}
\title{Read samrat parameters}
\usage{
samrat_read_params(param_file)
}
\arguments{
\item{param_file}{Name of parameter file}
}
\value{
Returns a list with 4 elements:
\itemize{
\item{"gen_pars"}{"General parameters \code{list}"}
\item{"var_pars"}{"Predictor variables \code{data.frame}"}
\item{"cf_pars"}{"Counterfactual variables \code{data.frame}}
}
}
\description{
Reads parameters from formatted \code{samrat} parameters file.
TODO: Link here to correct format for a parameter file
}
\details{
\code{samrat_read_params} Read samrat parameters
}
\examples{
{

param_file <- system.file(
"extdata/som_analysis_parameters.xlsx", package="samrat"
)
pars <- samrat_read_params(param_file)
names(pars)

}
}
