% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{samrat_read_predictors}
\alias{samrat_read_predictors}
\title{Read samrat predictor data}
\usage{
samrat_read_predictors(predictors_file, pars_list, check = TRUE)
}
\arguments{
\item{predictors_file}{Name of predictors file}

\item{pars_list}{Parameter outputs from \code{\link[=samrat_read_params]{samrat_read_params()}}}

\item{check}{Boolean for whether to check the output. Default = \code{TRUE}}
}
\value{
Returns a list with 3 elements:
\itemize{
\item{"demog_pars"}{"Demography parameters \code{list}"}
\item{"pop_sources"}{"Population source \code{list} of \code{data.frames}"}
\item{"dictionary"}{"Demography Dictionary \code{data.frame}"}
}
}
\description{
Reads predictors from formatted \code{samrat} predictors file.
TODO: Link here to correct format for a predictors file
}
\details{
\code{samrat_read_predictors} Read samrat predictors
}
\examples{
{

param_file <- system.file(
"extdata/som_analysis_parameters.xlsx", package="samrat"
)
pars_list <- samrat_read_params(param_file)

predictors_file <- system.file(
"extdata/som_predictor_data.xlsx", package="samrat"
)
predictors_list <- samrat_read_predictors(predictors_file, pars_list)
names(predictors_list)

}
}
