% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{samrat_read_strata}
\alias{samrat_read_strata}
\title{Read samrat strata parameters}
\usage{
samrat_read_strata(strata_file, pars_list)
}
\arguments{
\item{strata_file}{Name of strata file}

\item{pars_list}{Parameter outputs from \link{samrat_read_params}}
}
\value{
Returns a \code{data.frame} with 3 elements:
\itemize{
\item{"admin0"}{"Admin Level 0"}
\item{"admin1"}{"Admin Level 1"}
\item{"strata"}{"Stratum level"}
}
}
\description{
Reads parameters from formatted \code{samrat} strata file.
TODO: Link here to correct format for a parameter file
}
\details{
\code{samrat_read_strata} Read samrat strata parameters
}
\examples{
{

param_file <- system.file(
"extdata/som_analysis_parameters.xlsx", package="samrat"
)
pars_list <- samrat_read_params(param_file)

strata_file <- system.file(
"extdata/som_analysis_strata.xlsx", package="samrat"
)
strata <- samrat_read_strata(strata_file, pars_list)
names(strata)

}
}
