% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{samrat_read_surveymeta}
\alias{samrat_read_surveymeta}
\title{Read samrat survey metadata}
\usage{
samrat_read_surveymeta(surveymeta_file, pars)
}
\arguments{
\item{surveymeta_file}{Name of survey meta file}

\item{pars}{Parameter outputs from \link{samrat_read_params}}
}
\value{
Returns a list with 4 elements:
\itemize{
\item{"gen_pars"}{"General parameters \code{list}"}
\item{"var_pars"}{"Predictor variables \code{data.frame}"}
\item{"cf_pars"}{"Counterfactual variables \code{data.frame}}
}
}
\description{
Reads survey metadata from formatted \code{samrat} survey meta file.
TODO: Link here to correct format for a parameter file
}
\details{
\code{samrat_read_surveymeta} Read samrat survey metadata
}
\examples{
{

param_file <- system.file(
"extdata/som_analysis_parameters.xlsx", package="samrat"
)
pars <- samrat_read_params(param_file)

surveymeta_file <- system.file(
"extdata/som_survey_metadata.xlsx", package="samrat"
)
strata <- samrat_read_strata(surveymeta_file, pars)

}
}
