% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{samrat_timeseries}
\alias{samrat_timeseries}
\title{Create timeseries for analysis}
\usage{
samrat_timeseries(pars_list, strata)
}
\arguments{
\item{pars_list}{Parameter outputs from \code{\link[=samrat_read_params]{samrat_read_params()}}}

\item{strata}{Strata output from \code{\link[=samrat_read_strata]{samrat_read_strata()}}}
}
\value{
Returns a list of timeseries objects including:
\itemize{
\item{"ts"}{"Time series \code{data.frame}"}
\item{"tm_list"}{"Time parameters \code{list}"}
}
}
\description{
Uses parameters and strata to generate timeseries for mortality
inference.
}
\details{
\code{samrat_timeseries} Create timeseries for analysis in samrat
}
\examples{
{
  param_file <- system.file(
    "extdata/som_analysis_parameters.xlsx",
    package = "samrat"
  )
  pars_list <- samrat_read_params(param_file)

  strata_file <- system.file(
    "extdata/som_analysis_strata.xlsx",
    package = "samrat"
  )
  strata <- samrat_read_strata(strata_file, pars_list)

  tm_list <- samrat_timeseries(pars_list, strata)
  names(tm_list)
}
}
