% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_samrat_file}
\alias{read_samrat_file}
\title{read_samrat_file}
\usage{
read_samrat_file(name, sheet = NULL)
}
\arguments{
\item{name}{the name of a file within the inst/extdata folder.}
}
\description{
Load a file from within the inst/extdata folder of the
ICMDMM package. File extension must be one of .csv, .rds, or .xlsx
}
\details{
\code{read_file} read samrat package file
}
