% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{samrat_read_demography}
\alias{samrat_read_demography}
\title{Read samrat demographic data}
\usage{
samrat_read_demography(demog_file, pars)
}
\arguments{
\item{demog_file}{Name of demogrpahy file}
}
\value{
Returns a list with 2 elements:
\itemize{
\item{"demog_pars"}{"Demography parameters \code{list}"}
\item{"pop_sources"}{"Population source \code{list} of \code{data.frames}"}
}
}
\description{
Reads demography from formatted \code{samrat} demog_pars file.
TODO: Link here to correct format for a parameter file
}
\details{
\code{samrat_read_demography} Read samrat demography data
}
\examples{
{

demog_file <- system.file(
"extdata/som_demog_data.xlsx", package="samrat"
)
demog_pars <- samrat_read_demography(param_file)
names(demog_pars)

}
}
