% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{samrat_read_demography}
\alias{samrat_read_demography}
\title{Read samrat demographic data}
\usage{
samrat_read_demography(demog_file, pars_list, check = TRUE)
}
\arguments{
\item{demog_file}{Name of demogrpahy file}

\item{pars_list}{Parameter outputs from \link{samrat_read_params}}

\item{check}{Boolean for whether to check the output. Default = \code{TRUE}}
}
\value{
Returns a list with 5 elements:
\itemize{
\item{"demog_pars_list"}{"Demography parameters \code{list}"}
\item{"pop_sources_list"}{"Population source \code{list} of \code{data.frames}"}
\item{"demog_pars"}{"Demography parameters \code{data.frame}"}
\item{"pop_sources"}{"Population sources \code{data.frame}"}
\item{"dictionary"}{"Demography Dictionary \code{data.frame}"}
}
}
\description{
Reads demography from formatted \code{samrat} demog_pars file.
TODO: Link here to correct format for a parameter file
}
\details{
\code{samrat_read_demography} Read samrat demography data
}
\examples{
{


}
}
