% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{samrat_read_surveymeta}
\alias{samrat_read_surveymeta}
\title{Read samrat survey metadata}
\usage{
samrat_read_surveymeta(surveymeta_file, pars_list)
}
\arguments{
\item{surveymeta_file}{Name of survey meta file}

\item{pars_list}{Parameter outputs from \code{\link[=samrat_read_params]{samrat_read_params()}}}
}
\value{
Returns a data.frame with metadata on surveys available
}
\description{
Reads survey metadata from formatted \code{samrat} survey meta file.
TODO: Link here to correct format for a parameter file
}
\details{
\code{samrat_read_surveymeta} Read samrat survey metadata
}
\examples{
{

param_file <- system.file(
"extdata/som_analysis_parameters.xlsx", package = "samrat"
)
pars_list <- samrat_read_params(param_file)

surveymeta_file <- system.file(
"extdata/som_survey_metadata.xlsx", package = "samrat"
)
strata <- samrat_read_surveymeta(surveymeta_file, pars_list)

}
}
